### 2025-09-03 #8925 @editorialbot
[PRE REVIEW]: TranCIT: Transient Causal Interaction Toolbox
**Submitting author:** @sa-nouri (<a href="http://orcid.org/0000-0002-8846-9318">Salar Nouri</a>)
**Repository:** https://github.com/CMC-lab/TranCIT
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.12
