### 2025-09-05 2510632
Reduce JOSS paper to under 1000 words and fix compilation issues
- Condensed Summary section while preserving main ideas
- Streamlined Statement of need section
- Simplified Functionality section with concise bullet points
- Shortened Example section
- Condensed Implementation details
- Fixed YAML frontmatter formatting issues
- Resolved author normalization errors
- Paper now compiles successfully with all citations resolved
- Final word count: 965 words (under 1000 limit)
