+++ paper/paper.md
@@ -1,15 +1,10 @@
 ---
-# JOSS Paper Draft - Revised Version
-# the final release version/date, and DOI placeholders are filled in.
-
 title: "TranCIT: Transient Causal Interaction Toolbox"
 tags:
   - Python
   - neuroscience
   - causal inference
   - time series analysis
-  - Granger causality
-  - transfer entropy
   - Local field potential (LFP)
   - Electroencephalogram (EEG)
   - Magnetoencephalography (MEG)
@@ -19,7 +14,7 @@ authors:
     orcid: 0000-0002-8846-9318
     corresponding: true
   - name: Kaidi Shao
-    affiliation: "2"
+    affiliation: 2
     orcid: 0000-0002-3027-0090
     corresponding: true
   - name: Shervin Safavi
@@ -42,71 +37,63 @@ affiliations:
 
 date: "2025-08-30"
 bibliography: paper.bib
-csl: ieee.csl
 repository: "https://github.com/cmc-lab/trancit"
 crossref: true
-doi: "10.5281/zenodo.16998396" # Zenodo DOI for the archived version
+doi: "10.5281/zenodo.16998396"
 url: "https://trancit.readthedocs.io/en/latest/" 
 
 ---
 
 ## Summary
 
-The study of complex systems, e.g., neural circuits and cognitive functions, often requires understanding causal interactions that occur during brief, transient events [@logothetisHippocampalCorticalInteraction2012; @womelsdorfBurstFiringSynchronizes2014; @nitzanBrainwideInteractionsHippocampal2022; @safaviBrainComplexSystem2022; @safaviUncoveringOrganizationNeural2023; @lundqvistBetaBurstsCognition2024]. Traditional methods for estimating causality, such as Granger causality [@granger1969investigating] and Transfer Entropy (TE) [@Schreiber2000], are frequently challenged by the short-duration, non-stationary nature of these phenomena [@mitra2007observed]. Such methods typically assume stationarity and require long data segments, making them suboptimal for event-driven analysis.
+The study of complex systems, particularly neural circuits and cognitive functions, requires understanding causal interactions during brief, transient events [@logothetisHippocampalCorticalInteraction2012; @womelsdorfBurstFiringSynchronizes2014; @nitzanBrainwideInteractionsHippocampal2022; @safaviBrainComplexSystem2022; @safaviUncoveringOrganizationNeural2023; @lundqvistBetaBurstsCognition2024]. Traditional causality methods, such as Granger causality (GC) [@granger1969investigating] and Transfer Entropy (TE) [@Schreiber2000], assume stationarity and require long data segments, making them suboptimal for event-driven analysis [@mitra2007observed].
 
-We present `trancit` (Transient Causal Interaction Toolbox), a robust, open-source Python package implementing advanced causal inference methods specifically designed to address this challenge [@nouri_2025_trancit_package; @nouri2025trancit]. `trancit` provides a comprehensive pipeline for dynamic causal inference on multivariate time-series data. It is a Python-native implementation and extension of a powerful causal learning algorithm originally introduced in MATLAB [@shao2023transient]. By leveraging foundational Python libraries like NumPy [@harris2020array] and SciPy [@virtanen2020fundamental], `trancit` integrates seamlessly into modern data science and research workflows, making these advanced techniques accessible to a broader scientific community.
[…]
