### 2025-09-05 8d12a44
Update documentation references from DCS to TranCIT
- Fix citation section title to reference TranCIT instead of DCS
- Update license section to reference TranCIT package name
- Ensure consistent branding throughout documentation
