### 2025-10-25 #8925 @ankurankan
@crvernon Looks interesting, would be glad to help with the review.
