### 2025-11-01 CMC-lab/TranCIT#3 @markean
[DOCS] JOSS - Include code to reproduce paper simulation results
This is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/9302#issuecomment-3453561710.

## Documentation Issue Type

- [ ] Missing documentation
- [ ] Unclear explanation
- [ ] Incorrect information
- [ ] Broken links
- [ ] Missing examples
- [ ] API reference incomplete
- [ ] Tutorial needs improvement

## Location

**Where is the documentation issue?**

- [ ] README.md
- [ ] ReadTheDocs (`https://trancit.readthedocs.io`)
- [ ] API docstrings
- [ ] Examples directory
- [ ] Jupyter notebook tutorials
- [x] Other: JOSS submission paper

**Specific page/section**: [e.g., "DCS Calculator API reference", "Quickstart Tutorial - Step 3"]

## Current State

The paper includes an *Example* section presenting simulation results and figures comparing the method with those from Shao et al. (2023).

## Issue Description

The code and scripts used to reproduce the results and figures presented in the paper are not referenced or linked. Providing access to this code (e.g., in `paper/` or its subdirectory, or as a script/notebook) would enhance transparency and reproducibility.

## Suggested Improvement

**What would make this clearer?**

- [x] Add code examples
- [ ] Include mathematical formulations
- [ ] Provide research context/citations
- [ ] Add parameter descriptions
- [ ] Include output format explanation
- [ ] Add troubleshooting section
- [ ] Other: ________________

## Scientific Context

**What type of user would benefit from this improvement?**

- [ ] New users to causality analysis
- [ ] Experienced researchers
- [ ] Computational neuroscientists
- [ ] Financial analysts
- [ ] Climate scientists
- [ ] Software developers
- [ ] Students/educators

## Example Request

**If requesting examples, what would be most helpful?**

- [ ] Basic usage example
- [ ] Real-world data analysis
- [ ] Comparison between methods (DCS vs Granger vs TE)
- [ ] Performance benchmarking
- [ ] Statistical validation workflow
- [ ] Visualization examples

## Additional Context

- Is this blocking your research/work?
- Are there external resources that do this well?
- Would you be willing to contribute to the documentation?

## Priority

- [ ] Low - Minor improvement
- [ ] Medium - Would help many users
- [ ] High - Critical for package adoption
- [ ] Urgent - Blocking current users
