### 2025-11-01 CMC-lab/TranCIT#4 @markean
[DOCS] JOSS - Extra API Reference subsection found in docs
This is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/9302#issuecomment-3453561710.

## Documentation Issue Type

- [ ] Missing documentation
- [ ] Unclear explanation
- [ ] Incorrect information
- [ ] Broken links
- [ ] Missing examples
- [x] API reference incomplete
- [ ] Tutorial needs improvement

## Location

**Where is the documentation issue?**

- [ ] README.md
- [x] ReadTheDocs (`https://trancit.readthedocs.io`)
- [ ] API docstrings
- [ ] Examples directory
- [ ] Jupyter notebook tutorials
- [ ] Other: ________________

**Specific page/section**: [e.g., "DCS Calculator API reference", "Quickstart Tutorial - Step 3"]

## Current State

https://trancit.readthedocs.io/en/latest/api.html#id34.

https://trancit.readthedocs.io/en/latest/api.html#.

## Issue Description

An additional [API Reference subsection](https://trancit.readthedocs.io/en/latest/api.html#id34) appears under the main [API Reference page](https://trancit.readthedocs.io/en/latest/api.html#), which may be redundant or incorrectly structured.

## Suggested Improvement

Refactor, rename, or remove the duplicate subsection.

## Scientific Context

**What type of user would benefit from this improvement?**

- [ ] New users to causality analysis
- [ ] Experienced researchers
- [ ] Computational neuroscientists
- [ ] Financial analysts
- [ ] Climate scientists
- [ ] Software developers
- [ ] Students/educators

## Example Request

**If requesting examples, what would be most helpful?**

- [ ] Basic usage example
- [ ] Real-world data analysis
- [ ] Comparison between methods (DCS vs Granger vs TE)
- [ ] Performance benchmarking
- [ ] Statistical validation workflow
- [ ] Visualization examples

## Additional Context

- Is this blocking your research/work?
- Are there external resources that do this well?
- Would you be willing to contribute to the documentation?

## Priority

- [ ] Low - Minor improvement
- [ ] Medium - Would help many users
- [ ] High - Critical for package adoption
- [ ] Urgent - Blocking current users
