### 2025-11-01 CMC-lab/TranCIT#5 @markean
[DOCS] JOSS - Paper Acknowledgments uses undefined acronyms
This is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/9302#issuecomment-3453561710.

## Documentation Issue Type

- [ ] Missing documentation
- [x] Unclear explanation
- [ ] Incorrect information
- [ ] Broken links
- [ ] Missing examples
- [ ] API reference incomplete
- [ ] Tutorial needs improvement

## Location

**Where is the documentation issue?**

- [ ] README.md
- [ ] ReadTheDocs (`https://trancit.readthedocs.io`)
- [ ] API docstrings
- [ ] Examples directory
- [ ] Jupyter notebook tutorials
- [x] Other: JOSS submission paper

**Specific page/section**: [e.g., "DCS Calculator API reference", "Quickstart Tutorial - Step 3"]

## Current State

In the *Acknowledgments* section, the acronyms "KS" and "SS" are used to refer to individuals.

## Issue Description

These acronyms are not introduced in the paper, though those individuals can be inferred from the context.

## Suggested Improvement

Use full names instead of acronyms, as the paper is short and the names appear only a few times.

## Scientific Context

**What type of user would benefit from this improvement?**

- [ ] New users to causality analysis
- [ ] Experienced researchers
- [ ] Computational neuroscientists
- [ ] Financial analysts
- [ ] Climate scientists
- [ ] Software developers
- [ ] Students/educators

## Example Request

**If requesting examples, what would be most helpful?**

- [ ] Basic usage example
- [ ] Real-world data analysis
- [ ] Comparison between methods (DCS vs Granger vs TE)
- [ ] Performance benchmarking
- [ ] Statistical validation workflow
- [ ] Visualization examples

## Additional Context

- Is this blocking your research/work?
- Are there external resources that do this well?
- Would you be willing to contribute to the documentation?

## Priority

- [ ] Low - Minor improvement
- [ ] Medium - Would help many users
- [ ] High - Critical for package adoption
- [ ] Urgent - Blocking current users
