### 2025-11-01 CMC-lab/TranCIT#6 @markean
[DOCS] JOSS - API documentation examples cannot be executed independently
This is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/9302#issuecomment-3453561710.

## Documentation Issue Type

- [ ] Missing documentation
- [ ] Unclear explanation
- [ ] Incorrect information
- [ ] Broken links
- [ ] Missing examples
- [ ] API reference incomplete
- [x] Tutorial needs improvement

## Location

**Where is the documentation issue?**

- [ ] README.md
- [x] ReadTheDocs (`https://trancit.readthedocs.io`)
- [x] API docstrings
- [ ] Examples directory
- [ ] Jupyter notebook tutorials
- [ ] Other: ________________

**Specific page/section**: [e.g., "DCS Calculator API reference", "Quickstart Tutorial - Step 3"]

## Current State

Many example code blocks for class instantiation and method usage cannot be run independently. Some examples rely on objects defined elsewhere or in previous sections.

For instance, the example code in https://trancit.readthedocs.io/en/latest/api.html#trancit.core.base.BaseResult cannot be executed because the required objects `MyResult`, `cs_array`, and `te_array` are not defiend.

## Issue Description

End users often visit the API documentation to quickly understand how to use a class or method without reading the full guide. Currently, many examples cannot be executed independently, requiring users to refer to other sections.

## Suggested Improvement

Provide end-to-end, self-contained code for each class or method example. While this may introduce some duplication or boilerplate, it will improve usability and completeness of the documentation.

## Scientific Context

**What type of user would benefit from this improvement?**

- [ ] New users to causality analysis
- [ ] Experienced researchers
- [ ] Computational neuroscientists
- [ ] Financial analysts
- [ ] Climate scientists
- [ ] Software developers
- [ ] Students/educators

## Example Request

**If requesting examples, what would be most helpful?**

- [ ] Basic usage example
- [ ] Real-world data analysis
- [ ] Comparison between methods (DCS vs Granger vs TE)
- [ ] Performance benchmarking
- [ ] Statistical validation workflow
- [ ] Visualization examples

## Additional Context

- Is this blocking your research/work?
- Are there external resources that do this well?
- Would you be willing to contribute to the documentation?

## Priority

- [ ] Low - Minor improvement
- [ ] Medium - Would help many users
- [ ] High - Critical for package adoption
- [ ] Urgent - Blocking current users
