### 2025-11-01 CMC-lab/TranCIT#7 @markean
[BUG] JOSS - Line too long (E501) linting errors during testing
This is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/9302#issuecomment-3453561710.

## Bug Description

Linting errors occur due to lines exceeding the maximum length of 88 characters as specified in `pyproject.toml`.

## Steps to Reproduce

1. Install the package with the `dev` optional dependencies.
2. Run the linter:
```bash
flake8 trancit/
```

## Expected Behavior

No linting errors should be reported.

## Actual Behavior

Multiple `E501: line too long` errors are reported.

## Error Messages

```
trancit/utils/signal.py:119:89: E501 line too long (97 > 88 characters)
```

## Environment Information

- **TranCIT version**: [e.g., 1.0.10]
- **Python version**: [e.g., 3.11.5]
- **Operating System**: [e.g., macOS 14.0, Ubuntu 22.04, Windows 11]
- **NumPy version**: [e.g., 1.24.3]
- **SciPy version**: [e.g., 1.10.1]

## Data Information

- **Data type**: [e.g., EEG, LFP, simulated signals]
- **Data shape**: [e.g., (channels, time_points, trials)]
- **Sampling rate**: [if applicable]
- **Analysis method**: [DCS, Transfer Entropy, Granger Causality, etc.]

## Additional Context

- Is this related to a specific scientific application?
- Are you following a particular research methodology?
- Any relevant literature or expected results?

## Checklist

- [ ] I have searched existing issues to avoid duplicates
- [ ] I have provided a minimal reproducible example
- [ ] I have included version information
- [ ] I have included data shape/type information
