### 2025-11-01 CMC-lab/TranCIT#8 @markean
[BUG] JOSS - README examples produce `TypeError`
This is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/9302#issuecomment-3453561710.

## Bug Description

Example code in `README.md` produces errors when executed.

- [Event-based analysis pipeline](https://github.com/CMC-lab/TranCIT/?tab=readme-ov-file#event-based-analysis-pipeline)
- [Model selection and validation](https://github.com/CMC-lab/TranCIT/?tab=readme-ov-file#model-selection-and-validation)

## Steps to Reproduce

1. Install the package.
2. Copy and execute the example code blocks in Python.

## Expected Behavior

The code runs without errors.

## Actual Behavior

Executing the examples produces `TypeError`s.

## Error Messages

- **Event-based analysis pipeline**:
```
TypeError: PipelineConfig.__init__() missing 2 required positional arguments: 'bic' and 'output'
```

- **Model selection and validation**:
```
TypeError: BICSelector.__init__() got an unexpected keyword argument 'max_order'
```

## Environment Information

- **TranCIT version**: 1.0.12.post2
- **Python version**: [e.g., 3.11.5]
- **Operating System**: [e.g., macOS 14.0, Ubuntu 22.04, Windows 11]
- **NumPy version**: [e.g., 1.24.3]
- **SciPy version**: [e.g., 1.10.1]

## Data Information

- **Data type**: [e.g., EEG, LFP, simulated signals]
- **Data shape**: [e.g., (channels, time_points, trials)]
- **Sampling rate**: [if applicable]
- **Analysis method**: [DCS, Transfer Entropy, Granger Causality, etc.]

## Additional Context

- Is this related to a specific scientific application?
- Are you following a particular research methodology?
- Any relevant literature or expected results?

## Checklist

- [ ] I have searched existing issues to avoid duplicates
- [ ] I have provided a minimal reproducible example
- [ ] I have included version information
- [ ] I have included data shape/type information
