### 2025-11-17 d57c50b
Fix README.md: Address E501 linting errors and improve code examples
- Split long import statement in Event-Based Analysis Pipeline example
- Break long description line into multiple lines
- Format long function call in VAR Model Estimation example
- Split long citation line for better readability
- Add error handling to pipeline example
- Fix missing required parameters in PipelineConfig

Addresses issue #7: Line too long (E501) linting errors during testing
