### 2025-11-17 f092f1b
Fix docs/api.rst: Remove duplicate API Reference subsection
- Remove redundant 'API Reference' subsection at end of file
- Make API documentation examples executable independently by adding
  missing imports and data generation
- Add explanatory comments for examples requiring pipeline-generated data
- Fix Transfer Entropy, Granger Causality, RelativeDCS, and
  time_varying_causality examples

Addresses issue #4: Extra API Reference subsection found in docs
Ref: https://github.com/CMC-lab/TranCIT/issues/4
