+++ docs/api.rst
@@ -206,6 +206,10 @@ Transfer Entropy
    .. code-block:: python
    
       from trancit.causality import TransferEntropyCalculator
+      import numpy as np
+      
+      # Generate sample data
+      data = np.random.randn(2, 500, 20)  # 2 vars, 500 time points, 20 trials
       
       calculator = TransferEntropyCalculator(model_order=3)
       result = calculator.analyze(data)
@@ -237,6 +241,10 @@ Granger Causality
    .. code-block:: python
    
       from trancit.causality import GrangerCausalityCalculator
+      import numpy as np
+      
+      # Generate sample data
+      data = np.random.randn(2, 500, 20)  # 2 vars, 500 time points, 20 trials
       
       calculator = GrangerCausalityCalculator(model_order=4)
       result = calculator.analyze(data)
@@ -274,6 +282,11 @@ Relative Dynamic Causal Strength (rDCS)
    .. code-block:: python
    
       from trancit.causality import RelativeDCSCalculator
+      import numpy as np
+      
+      # Note: RelativeDCSCalculator requires event data and statistics
+      # typically generated by the pipeline. This is a simplified example.
+      # In practice, you would use PipelineOrchestrator to generate these.
       
       calculator = RelativeDCSCalculator(
           model_order=3,
@@ -282,7 +295,7 @@ Relative Dynamic Causal Strength (rDCS)
       )
       
       # Requires event data and statistics (typically from pipeline)
-      result = calculator.analyze(event_data, event_stats)
+      # result = calculator.analyze(event_data, event_stats)
 
 .. autoclass:: RelativeDCSResult
    :members:
@@ -312,20 +325,25 @@ Utility Functions
    .. code-block:: python
    
       from trancit.causality import time_varying_causality
+      import numpy as np
       
-      causal_params = {
-          'ref_time': 25,
-          'estim_mode': 'OLS',
-          'morder': 3,
-          'diag_flag': False,
-          'old_version': False
-      }
+      # Note: This function requires event data and statistics
+      # typically generated by the pipeline. This is a simplified example.
+      # In practice, you would use PipelineOrchestrator to generate these.
       
-      causality_results = time_varying_causality(event_data, stats, causal_params)
-      
-      dcs_values = causality_results['DCS']
-      te_values = causality_results['TE']
-      rdcs_values = causality_results['rDCS']
+      # causal_params = {
+      #     'ref_time': 25,
+      #     'estim_mode': 'OLS',
+      #     'morder': 3,
+      #     'diag_flag': False,
+      #     'old_version': False
+      # }
+      # 
+      # causality_results = time_varying_causality(event_data, stats, causal_params)
+      # 
+      # dcs_values = causality_results['DCS']
+      # te_values = causality_results['TE']
+      # rdcs_values = causality_results['rDCS']
 
 ************************************
 Pipeline System (`trancit.pipeline`)
@@ -925,46 +943,6 @@ Event-Based Analysis Workflow
            plt.legend()
            plt.title('Event-Based Causality Analysis')
            plt.show()
-
-*******************
-API Reference
-*******************
-
-The TranCIT package provides a modern class-based API for causality analysis:
-
-Class-Based API
-===============
-
-.. code-block:: python
-
-   # Class-based approach
-   from trancit import DCSCalculator, PipelineOrchestrator
-   
-   # Causality analysis
-   calculator = DCSCalculator(model_order=4, time_mode="inhomo")
-   result = calculator.analyze(data)
-   
-   cs = result.causal_strength
-   te = result.transfer_entropy
-   gc = result.granger_causality
-   coeffs = result.coefficients
-   te_cov = result.te_residual_cov
-   
-   # Pipeline analysis
[…]
