### 2025-11-17 ef26118
Fix paper/paper.md: Define undefined acronyms in Acknowledgments
- Define 'KS' as 'Kaidi Shao (KS)' in Acknowledgments section
- Define 'SS' as 'Shervin Safavi (SS)' in Acknowledgments section
- Add reference to code for reproducing simulation results

Addresses issue #5: Paper Acknowledgments uses undefined acronyms
Ref: https://github.com/CMC-lab/TranCIT/issues/5
