### 2025-11-17 cf24c5d
Fix Issue #3: Add LFP example replication for CA3-CA1 causality analysis
- Rename 0_plot_example_signals_refined.ipynb to plot_ca3_ca1_example_signals.ipynb
  - Add markdown explanations describing the notebook's goal
  - Fix file paths and imports (use trancit.utils.signal)
  - Add data source reference to CRCNS HC-3 dataset

- Convert 3_ca3_ca1_dcs_verfication.ipynb to plot_ca3_ca1_dcs_results.py
  - Script saves figures as .svg and .pdf formats
  - Add data source reference to CRCNS HC-3 dataset
  - Include command-line arguments for flexibility

- Rename lfp_pipeline.py to compute_ca3_ca1_
[…]
