+++ examples/plot_ca3_ca1_dcs_results.py
@@ -1,0 +1,370 @@
+"""
+Plot CA3-CA1 DCS Verification Results
+
+This script loads saved causality analysis results from the CA3-CA1 LFP pipeline
+and generates publication-quality figures showing:
+1. Event waveforms aligned by CA3 and CA1 peaks
+2. Causality measures (TE, DCS, rDCS) for different alignment conditions
+
+The causality results should be generated by running the LFP pipeline script
+(see `compute_ca3_ca1_causality.py`) before running this visualization script.
+
+Usage:
+    python plot_ca3_ca1_dcs_results.py [--data_dir DATA_DIR] [--output_dir OUTPUT_DIR]
+
+Output:
+    - event_waveforms.{svg,pdf}: Event waveform plots
+    - causality_results.{svg,pdf}: Causality analysis results
+"""
+
+import argparse
+import os
+
+import matplotlib.pyplot as plt
+import numpy as np
+
+
+def fill_std_known(mean_temp, std_temp, t, color):
+    """Fill between mean ± std and plot mean line."""
+    plt.fill_between(t, mean_temp - std_temp, mean_temp + std_temp, color=color, alpha=0.2)
+    plt.plot(t, mean_temp, color=color)
+
+
+def load_causality_data(data_dir, sess_name, n_chpair, align_mode, causal_methods, num_bootstraps=5):
+    """
+    Load causality analysis results from saved .npz files.
+    
+    Parameters
+    ----------
+    data_dir : str
+        Directory containing the saved .npz files
+    sess_name : str
+        Session name used in file naming
+    n_chpair : int
+        Number of channel pairs
+    align_mode : list
+        List of alignment modes (e.g., ['ca3_peak', 'ca1_peak'])
+    causal_methods : list
+        List of causal methods (e.g., ['TE', 'DCS', 'rDCS'])
+    num_bootstraps : int
+        Number of bootstrap samples
+        
+    Returns
+    -------
+    CausalStructure : dict
+        Dictionary containing causality results
+    CausalStructure_btsp : dict
+        Dictionary containing bootstrap results
+    Yt_stats_all : dict
+        Dictionary containing signal statistics
+    """
+    # Initialize data structures
+    CausalStructure = {'OLS': {}}
+    CausalStructure_btsp = {'OLS': {}}
+    Yt_stats_all = {'mean': {}, 'Sigma': {}}
+    
+    for align in align_mode:
+        VarAlign = align.replace('-', '_')
+        CausalStructure['OLS'][VarAlign] = {}
+        CausalStructure_btsp['OLS'][VarAlign] = {}
+        
+        for causal_method in causal_methods:
+            VarCausalM = causal_method.replace('-', '_')
+            # Initialize with NaN - will be filled as data is loaded
+            CausalStructure['OLS'][VarAlign][VarCausalM] = np.full((n_chpair, 401, 2), np.nan)
+            CausalStructure_btsp['OLS'][VarAlign][VarCausalM] = np.full(
+                (n_chpair, 401, 2, num_bootstraps), np.nan
+            )
+    
+    # Load data for each channel pair
+    for n_ch in range(1, n_chpair + 1):
+        for align in align_mode:
+            # Load main causality results
+            filename = os.path.join(
+                data_dir, 
+                f'ca3_ca1_{sess_name}_chpair_{n_ch}_{align}_model_causality.npz'
+            )
+            
+            if not os.path.exists(filename):
+                print(f"Warning: File not found: {filename}")
+                continue
+                
+            data = np.load(filename, allow_pickle=True)
+            
[…]
