### 2025-11-17 0351997
Fix Issue #7: Address E501 linting errors in trancit/pipeline/stages.py
- Break long log messages across multiple lines to comply with E501
- Improves code readability and passes linting checks
