### 2025-11-17 51284c6
Fix Issue #6: Make all API documentation examples self-contained and executable
- Fix BaseResult example: Add MyResult class definition and data arrays (cs_array, te_array)
- Fix BaseAnalyzer example: Add complete MyResult class and sample data usage
- Add BaseConfig example: Demonstrate configuration object usage
- Fix PipelineOrchestrator example: Add signal generation with generate_signals
- Fix Custom Pipeline example: Add all required imports and data generation
- Fix VAREstimator example: Use correct estimate_var_coefficients method with proper parameters
- Fix BICSelector example: Use c
[…]
