+++ docs/api.rst
@@ -46,13 +46,27 @@ Base Classes
    
    .. code-block:: python
    
-      from trancit.core.base import BaseAnalyzer
+      from trancit.core.base import BaseAnalyzer, BaseResult
       import numpy as np
       
+      # Define a custom result class
+      class MyResult(BaseResult):
+          """Custom result class for demonstration."""
+          pass
+      
+      # Define a custom analyzer
       class MyAnalyzer(BaseAnalyzer):
           def analyze(self, data):
               # Your analysis implementation here
               return MyResult(result=data.sum())
+      
+      # Generate sample data
+      data = np.random.randn(2, 100, 10)  # 2 vars, 100 time points, 10 trials
+      
+      # Use the analyzer
+      analyzer = MyAnalyzer()
+      result = analyzer.analyze(data)
+      print(f"Analysis result: {result.result}")
 
 .. autoclass:: BaseResult
    :members:
@@ -70,7 +84,19 @@ Base Classes
    .. code-block:: python
    
       from trancit.core.base import BaseResult
+      import numpy as np
       
+      # Define a custom result class
+      class MyResult(BaseResult):
+          """Custom result class for demonstration."""
+          pass
+      
+      # Generate sample data arrays
+      n_time = 100
+      cs_array = np.random.randn(n_time, 2)  # Causal strength array
+      te_array = np.random.randn(n_time, 2)  # Transfer entropy array
+      
+      # Create result object
       result = MyResult(causal_strength=cs_array, transfer_entropy=te_array)
       print(result.causal_strength.shape)
       result_dict = result.to_dict()
@@ -80,6 +106,28 @@ Base Classes
    :show-inheritance:
    
    Base class for configuration objects. Provides parameter validation and default value handling.
+   
+   **Example:**
+   
+   .. code-block:: python
+   
+      from trancit.core.base import BaseConfig
+      
+      # Define a custom configuration class
+      class MyConfig(BaseConfig):
+          """Custom configuration class for demonstration."""
+          pass
+      
+      # Create configuration object
+      config = MyConfig(param1=10, param2="value", param3=True)
+      
+      # Access parameters
+      print(f"param1: {config.param1}")
+      print(f"param2: {config.param2}")
+      
+      # Convert to dictionary
+      config_dict = config.to_dict()
+      print(f"Config as dict: {config_dict}")
 
 Exception Classes
 =================
@@ -384,8 +432,18 @@ Pipeline Orchestration
    
    .. code-block:: python
    
-      from trancit import PipelineOrchestrator
+      from trancit import PipelineOrchestrator, generate_signals
       from trancit.config import PipelineConfig, PipelineOptions, DetectionParams, CausalParams
+      import numpy as np
+      
+      # Generate sample data
+      data, _, _ = generate_signals(T=800, Ntrial=15, h=0.1,
+                                    gamma1=0.4, gamma2=0.6,
+                                    Omega1=0.9, Omega2=1.3)
+      
+      # Prepare signals for pipeline
+      original_signal = np.mean(data, axis=2)  # Average across trials
+      detection_signal = original_signal * 1.5  # Amplified for detection
       
       # Configure pipeline
       config = PipelineConfig(
@@ -468,6 +526,36 @@ Individual pipeline stages can be used independently for custom workflows.
 .. code-block:: python
 
    from trancit.pipeline.stages import InputValidationStage, EventDetectionStage, CausalityAnalysisStage
+   from trancit import generate_signals
+   from trancit.config import PipelineConfig, PipelineOptions, DetectionParams, CausalParams
+   import numpy as np
+   
+   # Generate sample data
[…]
