### 2025-11-17 33b2c9e
Fix Issue #6 (comprehensive): Make ALL API examples fully self-contained and executable
- Fixed RelativeDCSCalculator example: Now uses synthetic event data and stats in correct format (fully executable)
- Fixed time_varying_causality example: Now uses synthetic data with proper structure (fully executable)
- Fixed Event-Based Analysis Workflow: Added missing numpy and matplotlib imports
- All examples now include complete imports, data generation, and can run independently
- No examples rely on objects defined elsewhere
- Every code block is end-to-end and self-contained
