+++ docs/api.rst
@@ -332,18 +332,41 @@ Relative Dynamic Causal Strength (rDCS)
       from trancit.causality import RelativeDCSCalculator
       import numpy as np
       
-      # Note: RelativeDCSCalculator requires event data and statistics
-      # typically generated by the pipeline. This is a simplified example.
-      # In practice, you would use PipelineOrchestrator to generate these.
+      # Generate synthetic event data and statistics
+      # In practice, these would come from PipelineOrchestrator
+      np.random.seed(42)
+      model_order = 3
+      nvar = 2
+      n_obs = 50
+      n_trials = 10
       
+      # Event data: shape (nvar * (model_order + 1), n_obs, n_trials)
+      event_data = np.random.randn(nvar * (model_order + 1), n_obs, n_trials)
+      
+      # Statistics dictionary with required structure
+      stats = {
+          "OLS": {
+              "At": np.random.randn(n_obs, nvar, nvar * model_order + 1),
+              "Sigma_Et": np.array([np.eye(nvar) * 0.1 for _ in range(n_obs)]),
+          },
+          "Sigma": np.random.randn(
+              n_obs, nvar * (model_order + 1), nvar * (model_order + 1)
+          ),
+          "mean": np.random.randn(nvar * (model_order + 1), n_obs),
+      }
+      
+      # Create rDCS calculator
       calculator = RelativeDCSCalculator(
-          model_order=3,
-          reference_time=50,  # Reference time index
+          model_order=model_order,
+          reference_time=25,  # Reference time index
           estimation_mode="OLS"
       )
       
-      # Requires event data and statistics (typically from pipeline)
-      # result = calculator.analyze(event_data, event_stats)
+      # Analyze with event data and statistics
+      result = calculator.analyze(event_data, stats)
+      
+      print(f"rDCS result shape: {result.relative_dynamic_causal_strength.shape}")
+      print(f"Mean rDCS X→Y: {result.relative_dynamic_causal_strength[:, 1].mean():.4f}")
 
 .. autoclass:: RelativeDCSResult
    :members:
@@ -375,23 +398,50 @@ Utility Functions
       from trancit.causality import time_varying_causality
       import numpy as np
       
-      # Note: This function requires event data and statistics
-      # typically generated by the pipeline. This is a simplified example.
-      # In practice, you would use PipelineOrchestrator to generate these.
+      # Generate synthetic event data and statistics
+      # In practice, these would come from PipelineOrchestrator
+      np.random.seed(42)
+      model_order = 3
+      nvar = 2
+      n_obs = 50
+      n_trials = 10
       
-      # causal_params = {
-      #     'ref_time': 25,
-      #     'estim_mode': 'OLS',
-      #     'morder': 3,
-      #     'diag_flag': False,
-      #     'old_version': False
-      # }
-      # 
-      # causality_results = time_varying_causality(event_data, stats, causal_params)
-      # 
-      # dcs_values = causality_results['DCS']
-      # te_values = causality_results['TE']
-      # rdcs_values = causality_results['rDCS']
+      # Event data: shape (nvar * (model_order + 1), n_obs, n_trials)
+      event_data = np.random.randn(nvar * (model_order + 1), n_obs, n_trials)
+      
+      # Statistics dictionary with required structure
+      stats = {
+          "OLS": {
+              "At": np.random.randn(n_obs, nvar, nvar * model_order + 1),
+              "Sigma_Et": np.array([np.eye(nvar) * 0.1 for _ in range(n_obs)]),
+          },
+          "Sigma": np.random.randn(
+              n_obs, nvar * (model_order + 1), nvar * (model_order + 1)
+          ),
[…]
