### 2025-11-17 93835ec
Fix Issue #7: Update flake8 max-line-length to 88 and fix signal.py E501 errors
- Updated flake8 max-line-length from 100 to 88 to match Black/isort configuration
- Fixed all E501 errors in trancit/utils/signal.py by breaking long lines
- This ensures consistency across all linting tools (Black, isort, flake8)
