### 2025-11-17 1a5fb9b
Fix Issue #7: Resolve all E501 line too long errors across codebase
- Fixed all 53 E501 errors across 14 files by breaking long lines appropriately
- Updated flake8 max-line-length to 88 to match Black/isort configuration
- Fixed errors in:
  - trancit/causality/rdcs.py (8 errors)
  - trancit/core/base.py (1 error)
  - trancit/models/*.py (13 errors)
  - trancit/simulation/*.py (3 errors)
  - trancit/utils/*.py (28 errors)
- All lines now comply with 88-character limit
- Code maintains readability while meeting linting standards
