### 2025-11-17 1a3137b
Fix remaining E501 errors in bic_selection.py and simulation/utils.py
- Fixed line length issues in bic_selection.py docstring
- Fixed commented code line in simulation/utils.py
- All files now fully compliant with 88-character limit
