+++ trancit/simulation/utils.py
@@ -26,9 +26,7 @@ def morlet(start: float, end: float, num_points: int) -> np.ndarray:
     sigma = 1
     # sigma = end / (2 * np.pi)  # Standard deviation for Gaussian
     wavelet = (
-        (1 / np.sqrt(sigma)) * np.exp(1j * w0 * t) * np.exp(
-            -(t**2) / (2 * (sigma**2))
-        )
+        (1 / np.sqrt(sigma)) * np.exp(1j * w0 * t) * np.exp(-(t**2) / (2 * (sigma**2)))
     )
     # wavelet = np.exp(1j * start * t) * np.exp(
     #     -t ** 2 / (2 * (end ** 2))
