### 2025-11-17 78ec2b0
Format plot_ca3_ca1_dcs_results.py with Black
- Applied Black formatting to comply with code style
- Changed single quotes to double quotes (Black default)
- Fixed line length and spacing issues
- Ensures CI/CD pipeline passes Black checks
