+++ examples/plot_ca3_ca1_dcs_results.py
@@ -26,14 +26,18 @@ import numpy as np
 
 def fill_std_known(mean_temp, std_temp, t, color):
     """Fill between mean ± std and plot mean line."""
-    plt.fill_between(t, mean_temp - std_temp, mean_temp + std_temp, color=color, alpha=0.2)
+    plt.fill_between(
+        t, mean_temp - std_temp, mean_temp + std_temp, color=color, alpha=0.2
+    )
     plt.plot(t, mean_temp, color=color)
 
 
-def load_causality_data(data_dir, sess_name, n_chpair, align_mode, causal_methods, num_bootstraps=5):
+def load_causality_data(
+    data_dir, sess_name, n_chpair, align_mode, causal_methods, num_bootstraps=5
+):
     """
     Load causality analysis results from saved .npz files.
-    
+
     Parameters
     ----------
     data_dir : str
@@ -48,7 +52,7 @@ def load_causality_data(data_dir, sess_name, n_chpair, align_mode, causal_method
         List of causal methods (e.g., ['TE', 'DCS', 'rDCS'])
     num_bootstraps : int
         Number of bootstrap samples
-        
+
     Returns
     -------
     CausalStructure : dict
@@ -59,82 +63,90 @@ def load_causality_data(data_dir, sess_name, n_chpair, align_mode, causal_method
         Dictionary containing signal statistics
     """
     # Initialize data structures
-    CausalStructure = {'OLS': {}}
-    CausalStructure_btsp = {'OLS': {}}
-    Yt_stats_all = {'mean': {}, 'Sigma': {}}
-    
+    CausalStructure = {"OLS": {}}
+    CausalStructure_btsp = {"OLS": {}}
+    Yt_stats_all = {"mean": {}, "Sigma": {}}
+
     for align in align_mode:
-        VarAlign = align.replace('-', '_')
-        CausalStructure['OLS'][VarAlign] = {}
-        CausalStructure_btsp['OLS'][VarAlign] = {}
-        
+        VarAlign = align.replace("-", "_")
+        CausalStructure["OLS"][VarAlign] = {}
+        CausalStructure_btsp["OLS"][VarAlign] = {}
+
         for causal_method in causal_methods:
-            VarCausalM = causal_method.replace('-', '_')
+            VarCausalM = causal_method.replace("-", "_")
             # Initialize with NaN - will be filled as data is loaded
-            CausalStructure['OLS'][VarAlign][VarCausalM] = np.full((n_chpair, 401, 2), np.nan)
-            CausalStructure_btsp['OLS'][VarAlign][VarCausalM] = np.full(
+            CausalStructure["OLS"][VarAlign][VarCausalM] = np.full(
+                (n_chpair, 401, 2), np.nan
+            )
+            CausalStructure_btsp["OLS"][VarAlign][VarCausalM] = np.full(
                 (n_chpair, 401, 2, num_bootstraps), np.nan
             )
-    
+
     # Load data for each channel pair
     for n_ch in range(1, n_chpair + 1):
         for align in align_mode:
             # Load main causality results
             filename = os.path.join(
-                data_dir, 
-                f'ca3_ca1_{sess_name}_chpair_{n_ch}_{align}_model_causality.npz'
+                data_dir,
+                f"ca3_ca1_{sess_name}_chpair_{n_ch}_{align}_model_causality.npz",
             )
-            
+
             if not os.path.exists(filename):
                 print(f"Warning: File not found: {filename}")
                 continue
-                
+
             data = np.load(filename, allow_pickle=True)
-            
+
[…]
