### 2025-11-18 #9302 @sa-nouri
@crvernon @markean

We have revised our manuscript and the project based on the reviewer’s comments. We extend our gratitude to both the editor and the reviewer for their invaluable feedback, which has significantly enhanced the clarity and quality of our work.
We hope that this revised version satisfactorily addresses the reviewer’s concerns. Below are detailed, point-by-point responses to the reviewer's comments.

---

# JOSS Review Response

This document addresses all reviewer comments from the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9302#issuecomment-3453561710).

## Summary

All issues raised in the JOSS review have been addressed. A Pull Request has been created with all fixes. The changes maintain backward compatibility and improve code quality, documentation, and usability.

**Pull Request:** [https://github.com/CMC-lab/TranCIT/pull/new/joss-review-fixes](https://github.com/CMC-lab/TranCIT/pull/new/joss-review-fixes)

---

## Issue-by-Issue Response

### Issue #3: Replication of LFP Example
**Issue Link:** [https://github.com/CMC-lab/TranCIT/issues/3](https://github.com/CMC-lab/TranCIT/issues/3)
**Description:** The second example (LFP: CA3-CA1) in the DCS introduction lacked replication code.

**Changes Made:**
* **Renamed and enhanced example notebook:**
    * Added `examples/plot_ca3_ca1_example_signals.ipynb`
    * Fixed file paths to use relative paths with `os.path.join`
    * Corrected imports to use `trancit.utils.signal` instead of `dcs.utils.signal`
    * Added data source reference to CRCNS HC-3 dataset
* **Converted verification notebook to Python script:**
    * Added `examples/plot_ca3_ca1_dcs_results.py`
    * Script now saves figures as both `.svg` and `.pdf` formats
    * Added command-line arguments for flexibility
    * Added data source reference to CRCNS HC-3 dataset
* **Renamed pipeline script:**
    * Renamed `examples/lfp_pipeline.py` → `examples/compute_ca3_ca1_causality.py`
    * Already includes a data source reference to the CRCNS HC-3 dataset
* **Data Source:** All relevant files now mention that the data used is from the [CRCNS HC-3 dataset](https://crcns.org/data-sets/hc/hc-3/about-hc-3).

**Commits:**
* `cf24c5d` - Fix Issue #3: Add LFP example replication for CA3-CA1 causality analysis

### Issue #4: API Documentation - Duplicate Subsection
**Issue Link:** [https://github.com/CMC-lab/TranCIT/issues/4](https://github.com/CMC-lab/TranCIT/issues/4)
**Description:** The API documentation contained a duplicate "API Reference" subsection.

**Changes Made:**
* Removed the duplicate "API Reference" subsection from `docs/api.rst`
* Cleaned up the documentation structure

**Commits:**
* `f092f1b` - Fix docs/api.rst: Remove duplicate API Reference subsection

### Issue #5: Paper Acknowledgments - Undefined Acronyms
**Issue Link:** [https://github.com/CMC-lab/TranCIT/issues/5](https://github.com/CMC-lab/TranCIT/issues/5)
**Description:** The paper acknowledgments section used undefined acronyms (KS and SS).

**Changes Made:**
* Fixed undefined acronyms in `paper/paper.md`
* Changed "KS acknowledges" to "Kaidi Shao (KS) acknowledges"
* Changed "SS acknowledges" to "Shervin Safavi (SS) acknowledges"

**Commits:**
* `ef26118` - Fix paper/paper.md: Define undefined acronyms in Acknowledgments

### Issue #6: API Documentation Examples Cannot Be Executed Independently
**Issue Link:** [https://github.com/CMC-lab/TranCIT/issues/6](https://github.com/CMC-lab/TranCIT/issues/6)
**Description:** Many API documentation examples relied on objects defined elsewhere and could not be executed independently.

**Changes Made:**
* **BaseResult example:** Fixed the example that was specifically mentioned in the issue.
    * Added `MyResult` class definition.
    * Added `cs_array` and `te_array` data generation.
    * Example is now fully executable.
* **All other examples:** Made every example self-contained:
    * `BaseAnalyzer`: Added class definitions and usage.
    * `BaseConfig`: Added new example.
    * `RelativeDCSCalculator`: Converted from commented code to an executable example with synthetic data.
    * `time_varying_causality`: Converted from commented code to an executable example.
    * `PipelineOrchestrator`: Added signal generation.
    * `Custom Pipeline`: Added required imports and data.
    * `VAREstimator`: Fixed method calls and added data generation.
    * `BICSelector`: Fixed API usage and added a complete example.
    * `normalize_data`: Added data generation.
    * `plot_causality_results`: Added workflow.
    * `Event-Based Analysis Workflow`: Added missing imports (`numpy`, `matplotlib`).

**Result:** Every code example in the API documentation can now be copied and executed independently without requiring objects from other sections.

**Commits:**
* `51284c6` - Fix Issue #6: Make API documentation examples self-contained and executable
* `33b2c9e` - Fix Issue #6 (comprehensive): Make API examples fully self-contained and executable

### Issue #7: Line Too Long (E501) Linting Errors
**Issue Link:** [https://github.com/CMC-lab/TranCIT/issues/7](https://github.com/CMC-lab/TranCIT/issues/7)
**Description:** Multiple E501 linting errors due to lines exceeding 88 characters.

**Changes Made:**
* **Configuration update:**
    * Updated `pyproject.toml`: Changed `flake8` max-line-length from 100 to 88.
    * Ensures consistency with Black (88 characters) and isort (88 characters).
* **Fixed all E501 errors:** Resolved 53 errors across 14 files:
    * `trancit/utils/signal.py` (7 errors) - The file specifically mentioned in the issue
    * `trancit/causality/rdcs.py` (8 errors)
    * `trancit/core/base.py` (1 error)
    * `trancit/models/__init__.py` (1 error)
    * `trancit/models/bic_selection.py` (8 errors)
    * `trancit/models/model_selection.py` (1 error)
    * `trancit/models/var_estimation.py` (3 errors)
    * `trancit/simulation/__init__.py` (1 error)
    * `trancit/simulation/ar_simulation.py` (1 error)
    * `trancit/simulation/utils.py` (1 error)
    * `trancit/utils/core.py` (7 errors)
    * `trancit/utils/helpers.py` (5 errors)
    * `trancit/utils/plotting.py` (2 errors)
    * `trancit/utils/preprocess.py` (11 errors)
    * `trancit/utils/residuals.py` (3 errors)
    * `trancit/pipeline/stages.py` (12 errors)

**Verification:**
```bash
flake8 trancit/ --max-line-length=88
# Returns: 0 errors ✅
```

**Commits:**
* `d57c50b` - Fix README.md: Address E501 linting errors and improve code examples
* `0351997` - Fix Issue #7: Address E501 linting errors in trancit/pipeline/stages.py
* `93835ec` - Fix Issue #7: Update flake8 max-line-length to 88 and fix signal.py E501 errors
* `1a5fb9b` - Fix Issue #7: Resolve all E501 line too long errors across codebase
* `1a3137b` - Fix remaining E501 errors in bic_selection.py and simulation/utils.py

---

## Testing

All changes have been verified:
* ✅ **Pytest:** 78 tests passed, 8 skipped (intentional), 0 failures, 0 errors
* ✅ **Linting:** All E501 errors resolved, flake8 passes with 88-character limit
* ✅ **Code Quality:** Examples are executable and self-contained
* ✅ **Backward Compatibility:** All changes maintain existing functionality

**Test Command:**
```bash
pytest --tb=no -q
# Result: 78 passed, 8 skipped, 6 warnings
```

---

## Pull Request

* **Branch:** `joss-review-fixes`
* **Pull Request Link:** [https://github.com/CMC-lab/TranCIT/compare/main...joss-review-fixes](https://github.com/CMC-lab/TranCIT/compare/main...joss-review-fixes)
* **Total Commits:** 10 commits addressing all JOSS review issues

**Files Changed:**
* **Documentation:** `docs/api.rst`, `README.md`, `paper/paper.md`
* **Examples:** New/renamed files for LFP example replication
* **Code:** Fixed line length issues across 14 modules
* **Configuration:** Updated `pyproject.toml`

### Commit Summary

| Commit | Issue | Description |
| :--- | :--- | :--- |
| `d57c50b` | #7 | Fix `README.md`: Address E501 linting errors |
| `f092f1b` | #4 | Fix `docs/api.rst`: Remove duplicate API Reference subsection |
| `ef26118` | #5 | Fix `paper/paper.md`: Define undefined acronyms |
| `cf24c5d` | #3 | Fix Issue #3: Add LFP example replication |
| `0351997` | #7 | Fix Issue #7: Address E501 in `trancit/pipeline/stages.py` |
| `51284c6` | #6 | Fix Issue #6: Make API examples self-contained |
| `33b2c9e` | #6 | Fix Issue #6 : Make examples executable |
| `93835ec` | #7 | Fix Issue #7: Update flake8 config and fix `signal.py` |
| `1a5fb9b` | #7 | Fix Issue #7: Resolve all E501 errors across codebase |
| `1a3137b` | #7 | Fix remaining E501 errors in `bic_selection.py` and `utils.py` |
