### 2025-11-18 CMC-lab/TranCIT#10 @ankurankan
[FEATURE] JOSS - Amount of logging info
This is part of JOSS review: https://github.com/openjournals/joss-reviews/issues/9302

I tried the examples from the README, and the `Basic Causality Analysis` and `Event-Based Analysis Pipeline` generated a lot of logging messages. I was wondering if this was a deliberate design choice, as I think it would be challenging for users to follow or find relevant information in the amount of logging information being generated. Do you think it would be better to either increase the default logging level or reduce the amount of logging messages?
