### 2025-11-19 56c0b7e
Add logging verbosity control section to README
- Added 'Controlling Logging Verbosity' section after Quick Start examples
- Documents how to reduce INFO-level logging to WARNING or ERROR
- Explains how to increase verbosity to DEBUG for debugging
- Addresses reviewer feedback about excessive logging in examples
- Keeps examples clean while providing logging control documentation
