### 2025-11-19 c359572
Address reviewer feedback: expand Statement of need with specific package comparisons
- Add comparison with statsmodels and TransferEntropy packages
- Clarify that these packages offer standard methods (GC and TE) but lack event detection workflows
- Refine distinction between general-purpose discovery libraries and TranCIT's specialized transient analysis
- Emphasize TranCIT's unique features for non-stationary, event-related data
