### 2025-11-19 4356b0f
Address reviewer feedback: add comprehensive event detection preprocessing documentation
- Add detailed RST documentation (docs/event_detection_preprocessing.rst) covering:
  * All 5 preprocessing stages (Event Detection, Border Removal, Snapshot Extraction, Artifact Rejection, Statistics Computation)
  * Software architecture (Pipeline Design Pattern, Core Components, State Management)
  * API design (Configuration-Driven Architecture, High-Level and Low-Level APIs)
  * Implementation details and integration with causality analysis
- Integrate documentation into Sphinx docs (added to index.rs
[…]
