+++ docs/event_detection_preprocessing.rst
@@ -1,0 +1,369 @@
+.. _event_detection_preprocessing:
+
+#######################################
+Event Detection Preprocessing Pipeline
+#######################################
+
+Overview
+========
+
+TranCIT provides an integrated preprocessing pipeline for event detection, data alignment, and artifact rejection. This pipeline is specifically designed for analyzing transient, event-related neural dynamics where causal interactions occur during brief, intense bursts of activity (e.g., sharp wave-ripples, beta bursts, or other transient events).
+
+The preprocessing pipeline transforms raw continuous time-series data into aligned event trials, preparing the data for subsequent causality analysis using methods such as Dynamic Causal Strength (DCS) and relative Dynamic Causal Strength (rDCS).
+
+Preprocessing Stages
+====================
+
+The event detection preprocessing consists of five sequential stages:
+
+1. Event Detection
+------------------
+
+**Purpose:** Identifies transient events in the detection signal using a threshold-based approach.
+
+**Algorithm:**
+
+- Computes a detection threshold: ``threshold = mean(signal) + thres_ratio × std(signal)``
+- Identifies all time points where the detection signal exceeds this threshold
+- Applies one of two alignment methods to refine event locations:
+  
+  - **Peak alignment:** Refines detected locations to local peaks within a specified window, ensuring events are aligned to the maximum amplitude
+  - **Pooled alignment:** Uses detected locations directly, with optional location shrinking to reduce redundancy when events are detected in close temporal proximity
+
+**Configuration Parameters:**
+
+- ``thres_ratio`` (float): Multiplier for standard deviation in threshold calculation (higher values = fewer events detected)
+- ``align_type`` (str): Either ``'peak'`` or ``'pooled'`` alignment method
+- ``l_extract`` (int): Length of event windows to extract (used for peak alignment window)
+- ``shrink_flag`` (bool): Whether to apply location shrinking for pooled alignment
+- ``locs`` (Optional[np.ndarray]): Pre-provided event locations (if detection is disabled)
+
+**Output:** Array of event location indices in the original signal.
+
+2. Border Removal
+----------------
+
+**Purpose:** Filters out events that are too close to signal boundaries to ensure complete event windows can be extracted.
+
+**Algorithm:**
+
+- Removes event locations where ``location < l_extract`` or ``location > signal_length - l_extract``
+- Ensures that each event has sufficient data before and after its center point for complete window extraction
+
+**Configuration Parameters:**
+
+- ``l_extract`` (int): Minimum required window length (inherited from detection stage)
+
+**Output:** Filtered array of event locations with border events removed.
+
+3. Snapshot Extraction
+----------------------
+
+**Purpose:** Extracts fixed-length time windows around each aligned event location, creating a 3D array of event trials.
+
+**Algorithm:**
+
+- For each event location, extracts a window of length ``l_extract`` starting at offset ``l_start`` from the event center
+- Creates a 3D array of shape ``(n_variables × (model_order + 1), n_time_points, n_trials)``
+- Includes lagged variables up to ``model_order`` for VAR model estimation
+- Handles out-of-bounds windows by filling with NaN values
+
+**Configuration Parameters:**
+
+- ``l_extract`` (int): Length of each extracted event window
+- ``l_start`` (int): Offset from event center to start extraction (can be negative)
+- ``morder`` (int): Model order for VAR estimation (determines number of lagged variables)
+
+**Output:** 3D numpy array ``(n_variables × (model_order + 1), n_time_points, n_trials)`` containing aligned event snapshots.
+
+4. Artifact Rejection
+---------------------
+
[…]
