### 2025-11-19 cb251bb
Address reviewer feedback: revise Example section to focus on code usage
- Revise Example section to remove detailed experimental results discussion
- Add reference to examples/README.md for code snippets and figure reproduction instructions
- Add examples/README.md with detailed instructions for generating paper figures:
  * Figure 1: Using dcs_introduction.ipynb for simulated data analysis
  * Figure 2: Using compute_ca3_ca1_causality.py and plot_ca3_ca1_dcs_results.py for real neural data

This addresses the reviewer's concern that the Example section should focus on
demonstrating package u
[…]
