### 2025-11-19 b8264d2
Address reviewer feedback: add software architecture documentation
- Add comprehensive software_architecture.rst documenting:
  * Package structure (all 6 modules with design rationale)
  * Dependencies with detailed explanations
  * Design choices (configuration-driven, extensibility, error handling, reproducibility)
  * Testing and quality assurance
  * Performance considerations
- Remove redundant content between software_architecture.rst and event_detection_preprocessing.rst:
  * Pipeline architecture details moved to event_detection_preprocessing.rst with cross-reference
  * Performance c
[…]
