+++ docs/index.rst
@@ -201,6 +201,7 @@ Documentation
    examples
    api
    event_detection_preprocessing
+   software_architecture
    
 .. toctree::
    :maxdepth: 2
