### 2025-11-20 26c00a6
clarified 'causal inference' terms
