+++ paper/paper.md
@@ -48,13 +48,15 @@ url: "https://trancit.readthedocs.io/en/latest/"
 
 The study of complex systems, particularly neural circuits and cognitive functions, requires understanding causal interactions during brief, transient events [@logothetisHippocampalCorticalInteraction2012; @womelsdorfBurstFiringSynchronizes2014; @nitzanBrainwideInteractionsHippocampal2022; @safaviBrainComplexSystem2022; @safaviUncoveringOrganizationNeural2023; @lundqvistBetaBurstsCognition2024]. Traditional causality methods, such as Granger causality (GC) [@granger1969investigating] and Transfer Entropy (TE) [@Schreiber2000], assume stationarity and require long data segments, making them suboptimal for event-driven analysis [@mitra2007observed].
 
-We present `trancit` (Transient Causal Interaction Toolbox), an open-source Python package implementing advanced causal inference methods for transient dynamics [@nouri_2025_trancit_package; @nouri2025trancit]. `trancit` provides a comprehensive pipeline for dynamic causal inference on multivariate time-series data, extending a robust causal learning algorithm originally introduced in MATLAB [@shao2023transient]. Built on NumPy [@harris2020array] and SciPy [@virtanen2020fundamental], it integrates seamlessly into modern data science workflows.
+We present `trancit` (Transient Causal Interaction Toolbox), an open-source Python package for causal inference in multivariate time series, emphasising the quantification of directed causal interactions during transient dynamics[@nouri_2025_trancit_package; @nouri2025trancit]. `trancit` provides a comprehensive pipeline for dynamic causal analysis, extending a robust causal learning algorithm originally introduced in MATLAB [@shao2023transient]. Built on NumPy [@harris2020array] and SciPy [@virtanen2020fundamental], it integrates seamlessly into modern data science workflows.
 
 The package offers an integrated solution for end-to-end causal analysis, including:
 
-- **Advanced causal inference methods:** GC, TE, and robust Structural Causal Model(SCM)-based Dynamic Causal Strength (DCS) and relative Dynamic Causal Strength (rDCS)
-- **Event-based preprocessing:** Automated event detection, data alignment, and artifact rejection pipeline
+- **Advanced causal analysis methods:** GC, TE, and robust Structural Causal Model(SCM)-based Dynamic Causal Strength (DCS) and relative Dynamic Causal Strength (rDCS)
+- **Event-based preprocessing:** Automated event detection, data alignment, and artefact rejection pipeline
 - **Simulation tools:** Synthetic autoregressive (AR) time-series data generation with known causal structures for validation and exploring scenarios.
+
+TranCIT primarily estimates directed causal effects in multivariate time series. Because these effects are zero when no causation exists, testing against zero also enables causal discovery. However, our main focus is quantifying and tracking these effects over time, not proposing a discovery method.
 
 ## Statement of need
 
@@ -62,11 +64,11 @@ While many statistical methods focus on correlation, the ability to infer direct
 
 `trancit` bridges this gap with a fully open-source Python implementation. While packages like `statsmodels` [@seabold2010statsmodels] and `TransferEntropy` [@TransferEntropy_pkg] offer standard methods (GC and TE), and general-purpose libraries such as `causal-learn` [@zheng2024causal] and `tigramite` [@runge2022jakobrunge] focus on causal discovery, they all lack specialized features for analyzing transient, event-related data—specifically, integrated event detection and alignment workflows.
 
-Furthermore, unlike discovery libraries that often assume stationary dynamics, `trancit` provides a tailored solution that implements GC, TE, DCS, and rDCS, with configurations for non-stationary signals. This promotes reproducible research, lowers barriers to advanced causal inference, and supports applications in neuroscience, climatology, and economics.
[…]
