### 2025-11-20 2f0f98f
clarified 'causal inference' terms
