### 2025-11-20 472832b
Revert "clarified 'causal inference' terms"
This reverts commit 2f0f98f82b85c5a353d0f4e6a394335f365edf55.
