### 2025-11-28 50c2f2b
Address reviewer's minor comments: clarify MATLAB extension, add event detection details, expand error handling, fix end-to-end phrasing, standardize TranCIT/trancit usage, fix grammatical issues
