+++ paper/paper.md
@@ -48,44 +48,44 @@ url: "https://trancit.readthedocs.io/en/latest/"
 
 The study of complex systems, particularly neural circuits and cognitive functions, requires understanding causal interactions during brief, transient events [@logothetisHippocampalCorticalInteraction2012; @womelsdorfBurstFiringSynchronizes2014; @nitzanBrainwideInteractionsHippocampal2022; @safaviBrainComplexSystem2022; @safaviUncoveringOrganizationNeural2023; @lundqvistBetaBurstsCognition2024]. Traditional causality methods, such as Granger causality (GC) [@granger1969investigating] and Transfer Entropy (TE) [@Schreiber2000], assume stationarity and require long data segments, making them suboptimal for event-driven analysis [@mitra2007observed].
 
-We present `trancit` (Transient Causal Interaction Toolbox), an open-source Python package for causal inference in multivariate time series, emphasising the quantification of directed causal interactions during transient dynamics[@nouri_2025_trancit_package; @nouri2025trancit]. `trancit` provides a comprehensive pipeline for dynamic causal analysis, extending a robust causal learning algorithm originally introduced in MATLAB [@shao2023transient]. Built on NumPy [@harris2020array] and SciPy [@virtanen2020fundamental], it integrates seamlessly into modern data science workflows.
+We present `trancit` (Transient Causal Interaction Toolbox), an open-source Python package for causal inference in multivariate time series, emphasising the quantification of directed causal interactions during transient dynamics[@nouri_2025_trancit_package; @nouri2025trancit]. TranCIT provides a comprehensive pipeline for dynamic causal analysis, translating the robust causal learning algorithm originally introduced in MATLAB [@shao2023transient] to Python and extending it with a modular pipeline architecture, improved error handling, and enhanced integration with modern data science workflows. Built on NumPy [@harris2020array] and SciPy [@virtanen2020fundamental], TranCIT integrates seamlessly into modern data science workflows.
 
-The package offers an integrated solution for end-to-end causal analysis, including:
+The package offers an integrated solution for causal effect estimation and analysis, including:
 
-- **Advanced causal analysis methods:** GC, TE, and robust Structural Causal Model(SCM)-based Dynamic Causal Strength (DCS) and relative Dynamic Causal Strength (rDCS)
-- **Event-based preprocessing:** Automated event detection, data alignment, and artefact rejection pipeline
+- **Advanced causal analysis methods:** GC, TE, robust Structural Causal Model(SCM)-based Dynamic Causal Strength (DCS), and relative Dynamic Causal Strength (rDCS).
+- **Event-based preprocessing:** Automated event detection, data alignment, and artefact rejection pipeline.
 - **Simulation tools:** Synthetic autoregressive (AR) time-series data generation with known causal structures for validation and exploring scenarios.
 
-TranCIT primarily estimates directed causal effects in multivariate time series. Because these effects are zero when no causation exists, testing against zero also enables causal discovery. However, our main focus is quantifying and tracking these effects over time, not proposing a discovery method.
+TranCIT primarily estimates directed causal effects in multivariate time series. Because these effects are zero when no causation exists, testing against zero also enables causal discovery. However, our primary focus is quantifying and tracking these effects over time, not proposing a discovery method.
 
 ## Statement of need
 
 While many statistical methods focus on correlation, the ability to infer directed causal relationships offers deeper, more mechanistic insights into how complex systems function [@Seth2015]. A critical challenge is analyzing transient dynamics where interactions occur in brief, intense bursts. Existing methods are primarily implemented in proprietary software, such as MATLAB [@shao2023transient], which limits accessibility.
 
[…]
