### 2025-12-02 #9302 @sa-nouri
@ankurankan

## Response

Thank you for the feedback. We have addressed all the minor comments below:

### Comment I

I appreciate your suggestion. We have clarified that TranCIT translates the MATLAB algorithm to Python and extends it with additional features. Specifically, the revision now states:

> "translating the robust causal learning algorithm originally introduced in MATLAB [@shao2023transient] to Python and extending it with a modular pipeline architecture, improved error handling, and enhanced integration with modern data science workflows."

*(Revision in Summary section, line 51)*

### Comment II.A

Thanks for pointing out both shortcomings.

We have added a brief description of the event detection methods. The revision now states:

> "TranCIT provides integrated preprocessing for event detection using threshold-based methods with peak or pooled alignment, data alignment, and artifact rejection..."

More detailed information about the preprocessing pipeline, including all five stages (Event Detection, Border Removal, Snapshot Extraction, Artifact Rejection, and Statistics Computation), is available in the documentation at `docs/event_detection_preprocessing.rst` and the extended Arxiv version.

*(Revision in Causal estimation methods section, line 80)*

### Comment II.B

Thanks for pointing that out. We have revised it to reflect the description of our package better. In the revised version, we note:

> "The package offers an integrated solution for causal effect estimation and analysis".

We should also note that we have expanded the error-handling description to specify the types of errors it handles. The revision now states:

> "It includes error handling with custom exceptions for input validation, computation errors, configuration issues, data corruption, and numerical convergence problems..."

The custom exception hierarchy (`ValidationError`, `ComputationError`, `ConfigurationError`, `DataError`, `ConvergenceError`, `SingularMatrixError`) provides informative error messages that help users identify and resolve issues quickly.

*(Revision in Implementation details section, line 96)*

### Comment III

Thanks for pointing that out. We agree that "end-to-end" was inaccurate, as it implies coverage of all causal inference tasks (discovery, identification, estimation, etc.), which TranCIT does not provide. We have revised the phrasing to:

> "The package offers an integrated solution for causal effect estimation and analysis, including:"

*(Revision in Summary section, line 53)*

### Comment IV

Thanks for spotting this inconsistency. We have standardized the usage throughout the paper following the convention:

* **"TranCIT"** (capitalized): Used when writing about what the software does in normal English sentences.
* **"`trancit`"** (lowercase in backticks): Used when referring to the code artifact, package name, folder name, or installation commands.

*(Revisions throughout the paper)*

### Comment V

Apologies, we missed them. We have fixed all grammatical issues:

* **Double "and" in bullet point:** Changed from "GC, TE, and robust Structural Causal Model(SCM)-based Dynamic Causal Strength (DCS) and relative Dynamic Causal Strength (rDCS)" to "GC, TE, robust Structural Causal Model(SCM)-based Dynamic Causal Strength (DCS), and relative Dynamic Causal Strength (rDCS)".
* **Missing periods:** Added closing periods to all bullet points:
    * "Advanced causal analysis methods: ... (rDCS)."
    * "Event-based preprocessing: ... pipeline."
    * "Simulation tools: ... scenarios." (already had a period)

*(Revisions in Summary section, lines 55-57)*

---

All revisions have been made while respecting the word limit constraints. We believe these changes improve clarity and accuracy without significantly expanding the manuscript length.
