+++ paper/paper.bib
@@ -130,7 +130,7 @@
 }
 
 @article{virtanen2020fundamental,
-  title={Fundamental algorithms for scientific computing in python and SciPy 1.0 contributors. SciPy 1.0},
+  title={Fundamental algorithms for scientific computing in {P}ython and SciPy 1.0 contributors. SciPy 1.0},
   author={Virtanen, P and Gommers, R and Oliphant, TE and Haberland, M and Reddy, T and Cournapeau, D and Burovski, E and Peterson, P and Weckesser, W and Bright, J and others},
   journal={Nat. Methods},
   volume={17},
@@ -170,7 +170,7 @@
 }
 
 @article{zheng2024causal,
-  title={Causal-learn: Causal discovery in python},
+  title={Causal-learn: Causal discovery in {P}ython},
   author={Zheng, Yujia and Huang, Biwei and Chen, Wei and Ramsey, Joseph and Gong, Mingming and Cai, Ruichu and Shimizu, Shohei and Spirtes, Peter and Zhang, Kun},
   journal={Journal of Machine Learning Research},
   volume={25},
@@ -229,9 +229,6 @@ Subject\_term\_id: consolidation;hippocampus;spatial-memory},
   timestamp = {2022-07-01T18:22:17Z},
   file = {joo2018the.pdf:/home/ssafavi/Nextcloud/libraries/zoteroLib/neuroscience/states/events/computation/joo2018the.pdf:application/pdf}
 }
-% == BibTeX quality report for jooHippocampalSharpWave2018:
-% ? unused Journal abbreviation ("Nat Rev Neurosci")
-% ? unused Library catalog ("www.nature.com")
 
 @article{womelsdorfBurstFiringSynchronizes2014,
   title = {Burst Firing Synchronizes Prefrontal and Anterior Cingulate Cortex during Attentional Control},
@@ -325,8 +322,6 @@ Subject\_term\_id: consolidation;hippocampus;spatial-memory},
   groups = {events},
   timestamp = {2023-11-21T20:07:27Z}
 }
-% == BibTeX quality report for nitzanBrainwideInteractionsHippocampal2022:
-% ? unused Library catalog ("pnas.org (Atypon)")
 
 @article{safaviUnivariateMultivariateCoupling2021,
   title = {From {{Univariate}} to {{Multivariate Coupling}} between {{Continuous Signals}} and {{Point Processes}}: {{A Mathematical Framework}}},
@@ -346,9 +341,6 @@ Subject\_term\_id: consolidation;hippocampus;spatial-memory},
   timestamp = {2021-05-19T13:10:08Z},
   file = {safavi2021fro.pdf:/home/ssafavi/Nextcloud/libraries/zoteroLib/sspub/papers/safavi2021fro.pdf:application/pdf}
 }
-% == BibTeX quality report for safaviUnivariateMultivariateCoupling2021:
-% ? Title looks like it was stored in title-case in Zotero
-% ? unused Library catalog ("Silverchair")
 
 @article{liuECannulaRevealsAnatomical2022b,
   title = {E-{{Cannula}} Reveals Anatomical Diversity in Sharp-Wave Ripples as a Driver for the Recruitment of Distinct Hippocampal Assemblies},
@@ -368,8 +360,6 @@ Subject\_term\_id: consolidation;hippocampus;spatial-memory},
   groups = {@collection},
   timestamp = {2023-05-21T22:57:15Z}
 }
-% == BibTeX quality report for liuECannulaRevealsAnatomical2022b:
-% ? unused Library catalog ("www.cell.com")
 
 @article{weichwaldDistributionalRobustnessGuiding2020,
[…]
