### 2025-12-30 3547b6d
Fix: Event detection link
