### 2025-12-30 17862c1
Fix: Grammatical Issues in the paper.md
