+++ paper/paper.bib
@@ -465,7 +465,7 @@ Subject\_term\_id: consolidation;hippocampus;spatial-memory},
 
 @misc{TransferEntropy_pkg,
   author = {Goeminne, Rutger},
-  title = {TransferEntropy: A Python library for transfer entropy calculation},
+  title = {TransferEntropy: A {P}ython library for transfer entropy calculation},
   year = {2019},
   publisher = {GitHub},
   journal = {GitHub repository},
