### 2025-12-31 af4d4db
Update CITATION.cff with JOSS publication information and Zenodo DOI
