+++ CITATION.cff
@@ -1,36 +1,48 @@
-cff-version: 1.2.0
-title: "TranCIT: Transient Causal Interaction Toolbox"
-message: "If you use this software, please cite it as below."
+cff-version: "1.2.0"
 authors:
-  - given-names: Salar
-    family-names: Nouri
-    email: salr.nouri@gmail.com
+- family-names: Nouri
+  given-names: Salar
+  orcid: "https://orcid.org/0000-0002-8846-9318"
+- family-names: Shao
+  given-names: Kaidi
+  orcid: "https://orcid.org/0000-0002-3027-0090"
+- family-names: Safavi
+  given-names: Shervin
+  orcid: "https://orcid.org/0000-0002-2868-530X"
+contact:
+- family-names: Nouri
+  given-names: Salar
+  orcid: "https://orcid.org/0000-0002-8846-9318"
+- family-names: Shao
+  given-names: Kaidi
+  orcid: "https://orcid.org/0000-0002-3027-0090"
+- family-names: Safavi
+  given-names: Shervin
+  orcid: "https://orcid.org/0000-0002-2868-530X"
+doi: 10.5281/zenodo.18098917
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Nouri
+    given-names: Salar
     orcid: "https://orcid.org/0000-0002-8846-9318"
-    affiliation: "School of Electrical and Computer Engineering, College of Engineering, University of Tehran, Tehran, Iran"
-  - given-names: Kaidi
-    family-names: Shao
+  - family-names: Shao
+    given-names: Kaidi
     orcid: "https://orcid.org/0000-0002-3027-0090"
-    affiliation: "International Center for Primate Brain Research (ICPBR), Center for Excellence in Brain Science and Intelligence Technology (CEBSIT), Chinese Academy of Sciences (CAS), Shanghai, China"
-  - given-names: Shervin
-    family-names: Safavi
+  - family-names: Safavi
+    given-names: Shervin
     orcid: "https://orcid.org/0000-0002-2868-530X"
-    affiliation: "Computational Neuroscience, Department of Child and Adolescent Psychiatry, Faculty of Medicine, Technische Universität Dresden, Dresden 01307, Germany"
-repository-code: "https://github.com/CMC-lab/TranCIT"
-url: "https://trancit.readthedocs.io"
-license: BSD-2-Clause
-version: 1.0.0
-date-released: 2025-08-30
-doi: "10.5281/zenodo.16998396"
-keywords:
-  - causality
-  - time-series
-  - neuroscience
-  - statistics
-  - causal-inference
-  - trancit
-  - transfer-entropy
-  - granger-causality
-  - transient-neural-events
-  - local-field-potential
-  - electroencephalogram
-  - magnetoencephalography 
+  date-published: 2025-12-30
+  doi: 10.21105/joss.09302
+  issn: 2475-9066
+  issue: 116
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 9302
+  title: "TranCIT: Transient Causal Interaction Toolbox"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.09302"
+  volume: 10
+title: "TranCIT: Transient Causal Interaction Toolbox"
