### 2025-12-31 47b493f
Add JOSS paper badge and citation, update Zenodo DOI
